pyproject.toml
[tool.poetry]
name = "vectr-tools"
version = "0.1.0"
description = ""
authors = ["thebleucheese <30511683+thebleucheese@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pydantic = "^1.9.0"
gql = {extras = ["requests"], version = "^3.0.0"}
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
